# 2DAW/DWES/htdocs/EjerciciosBasicos/P2Trim/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Nosecaen API Documentation",
        "description": "API para la gestión de incidencias y tareas",
        "contact": {
            "email": "soporte@nosecaen.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor de Desarrollo"
        }
    ],
    "paths": {
        "/api/tasks": {
            "get": {
                "tags": [
                    "Tareas"
                ],
                "summary": "Obtener lista de tareas",
                "description": "Devuelve una lista de todas las tareas registradas",
                "operationId": "getTasksList",
                "responses": {
                    "200": {
                        "description": "Operación exitosa"
                    },
                    "401": {
                        "description": "No autorizado"
                    }
                }
            },
            "post": {
                "tags": [
                    "Tareas"
                ],
                "summary": "Crear nueva tarea",
                "description": "Crea una tarea en el sistema",
                "operationId": "storeTask",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "description",
                                    "contact_person"
                                ],
                                "properties": {
                                    "description": {
                                        "type": "string",
                                        "example": "Arreglar ascensor"
                                    },
                                    "contact_person": {
                                        "type": "string",
                                        "example": "Juan Pérez"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Creado correctamente"
                    }
                }
            }
        },
        "/api/tasks/{id}": {
            "get": {
                "tags": [
                    "Tareas"
                ],
                "summary": "Obtener una tarea específica",
                "operationId": "getTaskById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la tarea",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Exitosa"
                    },
                    "404": {
                        "description": "No encontrada"
                    }
                }
            },
            "put": {
                "tags": [
                    "Tareas"
                ],
                "summary": "Actualizar una tarea",
                "operationId": "updateTask",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la tarea",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "description": {
                                        "type": "string",
                                        "example": "Arreglar ascensor"
                                    },
                                    "contact_person": {
                                        "type": "string",
                                        "example": "Juan Pérez"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Actualizada correctamente"
                    },
                    "404": {
                        "description": "No encontrada"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Tareas"
                ],
                "summary": "Eliminar una tarea",
                "operationId": "deleteTask",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la tarea",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Eliminada correctamente"
                    },
                    "404": {
                        "description": "No encontrada"
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Tareas",
            "description": "Endpoints para la gestión de tareas de incidencias"
        }
    ]
}